ml-service/requirements.txt
fastapi
uvicorn
transformers
torch
scikit-learn
sentence-transformers
python-dotenv
numpy
pandas
mysql-connector-python
sqlalchemy
pydantic
httpx
joblib
